requirements.txt
# OCR
pytesseract>=0.3.10
Pillow>=10.0.0
pymupdf>=1.24.0
